Renamed 'DataStoreMock' -> 'datastoremock'
